update rotate_cube v1

+++ mani_skill/utils/structs/articulation.py
@@ -183,7 +183,7 @@ class Articulation(BaseStruct[physx.PhysxArticulation]):
         return torch.tensor(
             [px_articulation.gpu_index for px_articulation in self._objs],
             device=self.device,
-            dtype=torch.int32,
+            dtype=torch.long,
         )
 
     def get_state(self):
